+++ lib/service/StudentCourses.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
+import 'package:geolocator/geolocator.dart';
 import 'package:http/http.dart';
 import 'package:illinois/model/StudentCourse.dart';
 import 'package:illinois/service/Auth2.dart';
@@ -13,6 +14,7 @@ import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:rokwire_plugin/model/explore.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
+import 'package:rokwire_plugin/service/location_services.dart';
 import 'package:rokwire_plugin/service/network.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/service.dart';
@@ -26,6 +28,7 @@ class StudentCourses with Service implements NotificationsListener, ExploreJsonH
   static const String ExternalAuthorizationHeader = "External-Authorization";
   
   static const String _courseTermsName = "course.terms.json";
+  static const String _requireAdaSetting = 'edu.illinois.rokwire.settings.student_course.require_ada';
 
   late Directory _appDocDir;
   
@@ -184,9 +187,16 @@ class StudentCourses with Service implements NotificationsListener, ExploreJsonH
   Future<List<StudentCourse>?> loadCourses({required String termId}) async {
     if (StringUtils.isNotEmpty(Config().gatewayUrl) && StringUtils.isNotEmpty(termId) && StringUtils.isNotEmpty(Auth2().uin)) {
       String url = "${Config().gatewayUrl}/courses/studentcourses?id=${Auth2().uin}&termid=$termId";
+      Position? position = await _userLocation;
+      if (position != null) {
+        url += "&lat=${position.latitude}&long=${position.longitude}";
+        if (requireAda) {
+          url += "&adaOnly=true";
+        }
+      }
       Response? response = await Network().get(url, auth: Auth2(), headers: { ExternalAuthorizationHeader: Auth2().uiucToken?.accessToken });
       String? responseString = (response?.statusCode == 200) ? response?.body : null;
-      /* TMP: String? responseString = '''[
+      /* TMP String? responseString = '''[
         {"coursetitle":"Thesis Research","courseshortname":"TAM 599","coursenumber":"25667","instructionmethod":"IND","coursesection":{"days":"","meeting_dates_or_range":"08/22/2022 - 12/07/2022","room":"","buildingname":"","buildingid":"","instructiontype":"IND","instructor":"Johnson, Harley","start_time":"","endtime":"","building":{"ID":"","Name":"","Number":"","FullAddress":"","Address1":"","Address2":"","City":"","State":"","ZipCode":"","ImageURL":"","MailCode":"","Entrances":null,"Latitude":0,"Longitude":0}}},
         {"coursetitle":"Atomic Scale Simulations","courseshortname":"CSE 485","coursenumber":"64706","instructionmethod":"LCD","coursesection":{"days":"Tu,Th","meeting_dates_or_range":"08/22/2022 - 12/07/2022","room":"305","buildingname":"Materials Science & Eng Bld","buildingid":"0034","instructiontype":"LCD","instructor":"Wagner, Lucas","start_time":"0930","endtime":"0930","building":{"ID":"3bb21766-3ad4-47e0-a472-d3c6cdbb07d0","Name":"Materials Science and Engineering Building","Number":"","FullAddress":"1304 W Green St  Urbana, IL 61801","Address1":"1304 W Green St","Address2":"","City":"Urbana","State":"IL","ZipCode":"61801","ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultBuildingImage.png","MailCode":"","Entrances":[{"ID":"d0564da7-66d6-452e-9198-65c860a57594","Name":"matscience_north","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.11109,"Longitude":-88.22594},{"ID":"1c1e1e4a-33f1-416f-8ccc-c5f990f347f7","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.111046,"Longitude":-88.22625},{"ID":"7d41bd0d-7b62-44e4-88ff-c27b3845bcdd","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.11115,"Longitude":-88.22617},{"ID":"0e9046ab-eeeb-44e5-81c3-6bdd635a3b59","Name":"matscience_south","ADACompliant":true,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.11071,"Longitude":-88.22608},{"ID":"41e5ef8a-6ed3-4229-a8e7-8f234e670900","Name":"matscience_west","ADACompliant":true,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.11078,"Longitude":-88.22641},{"ID":"ef2e8b18-2130-4512-982f-3c77ebc833b8","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.11079,"Longitude":-88.22575}],"Latitude":40.1109,"Longitude":-88.22608}}},
         {"coursetitle":"Advanced Continuum Mechanics","courseshortname":"TAM 545","coursenumber":"39105","instructionmethod":"LCD","coursesection":{"days":"M,W","meeting_dates_or_range":"08/22/2022 - 12/07/2022","room":"1047","buildingname":"Sidney Lu Mech Engr Bldg","buildingid":"0112","instructiontype":"LCD","instructor":"Starzewski, Martin","start_time":"1300","endtime":"1300","building":{"ID":"f96efb1f-8973-40d4-b602-c0a680897ad3","Name":"Sidney Lu Mechanical Engineering Building","Number":"","FullAddress":"1206 W Green St  Urbana, IL 61801","Address1":"1206 W Green St","Address2":"","City":"Urbana","State":"IL","ZipCode":"61801","ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultBuildingImage.png","MailCode":"","Entrances":[{"ID":"2b5f7e7f-ad33-4bff-a132-1aec7248a68d","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.111217,"Longitude":-88.22483},{"ID":"a5c9fdaa-3527-4fe7-a606-cbe62d6b419d","Name":"mechengbldg_north","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.11129,"Longitude":-88.22523},{"ID":"322f1204-8f34-4bd6-aac5-a0fae69e971f","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.111195,"Longitude":-88.225334},{"ID":"52844aac-1e27-471b-b7df-ee34f8dfe8e0","Name":"mechengbldg_west","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.110783,"Longitude":-88.225296},{"ID":"612c26ba-71ac-4b7a-be61-2459711ba589","Name":"mechengbldg_south","ADACompliant":true,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.110718,"Longitude":-88.22477},{"ID":"fee42e83-fcce-4344-aa9c-6806a2584961","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.11088,"Longitude":-88.22482},{"ID":"f7516216-53dd-4abe-b607-fc0ac698f043","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.111053,"Longitude":-88.22482},{"ID":"8c2ff914-ee51-4666-b468-02f73d567029","Name":"","ADACompliant":false,"Available":true,"ImageURL":"https://images.ccf.virtual.illinois.edu/DefaultEntranceImage.jpg","Latitude":40.110867,"Longitude":-88.2241}],"Latitude":40.11095,"Longitude":-88.224884}}},
@@ -198,6 +208,16 @@ class StudentCourses with Service implements NotificationsListener, ExploreJsonH
     }
   }
 
+  // Settings
+
+  bool get requireAda => (Auth2().prefs?.getBoolSetting(_requireAdaSetting) == true);
+  set requireAda(bool value) => Auth2().prefs?.applySetting(_requireAdaSetting, value);
+
+  // User Location
+
+  Future<bool> get _userLocationEnabled async => Auth2().privacyMatch(2) && (await LocationServices().status == LocationServicesStatus.permissionAllowed);
+  Future<Position?> get _userLocation async => await _userLocationEnabled ? await LocationServices().location : null;
+    
   // ExploreJsonHandler
 
   @override bool exploreCanJson(Map<String, dynamic>? json) => StudentCourse.canJson(json);